Added new version of spell-checker.
Also fixed bugs in old version and added new method "update_vocab"
